+++ release/components/button/index.js
@@ -1,3 +1,4 @@
 export * from './button.module';
 export * from './file-button.component';
+export * from './button.component';
 //# sourceMappingURL=index.js.map