+++ release/components/index.js
@@ -23,4 +23,5 @@ export * from './split';
 export * from './hotkeys';
 export * from './nag';
 export * from './long-press';
+export * from './radiobutton';
 //# sourceMappingURL=index.js.map