+++ release/components/long-press/index.js
@@ -1,2 +1,3 @@
 export * from './long-press-button.module';
+export * from './long-press-button.component';
 //# sourceMappingURL=index.js.map