+++ release/components/split/index.js
@@ -1,2 +1,5 @@
 export * from './split.module';
+export * from './split.component';
+export * from './split-handle.component';
+export * from './split-area.directive';
 //# sourceMappingURL=index.js.map